Use metapost to create a item marker
I like to use filled square as the top-level item marker. Rather than using math
symbols for the item marker, I use Metapost to draw the square.

After using this style for a while, I prefered not to use bullets for first
level of itemize, so I set the bullet symbol to empty.

For more details on using MetaPost, read:

- http://wiki.contextgarden.net/MetaPost_in_ConTeXt
- http://www.pragma-ade.com/general/manuals/metafun-s.pdf

For details on symbols, read:

- http://wiki.contextgarden.net/Symbols