Set alignment
I do not like hyphenated text in presentations, so I set alignment to
`nothyphenated`. This is a tricky setting, as TeX has to work harder to justify
paragraphs. So, I also set a low value of tolerance to avoid overfull boxes.
This is not perfect, and in some cases, it is upto the author to insert `\crlf`
to manually break the lines.