Layer to add material at arbitrary locations
Sometimes, you need to add material at a specific location in a slide, for
example, an image on top right. In principle, one can use floats for such
material, but it can be hard to control the placement of floats. Therefore, I
define a layer to place such material.

See http://wiki.contextgarden.net/Layers for details on Layers.

In addition to the usual x,y coordinates to place material on a layer, one can
also use `line` and `column` to place material on a layer. The `column` argument
is really useful as it leads to consistent locations. See the details manual
for, well, details.